[arquillian/arquillian-extension-liferay/issues/8] Move code from arquillian-cotainer-liferay to the new extension
