+++ arquillian-container-osgi-with-extensions/src/test/java/test/org/arquillian/container/osgi/remote/deploy/processor/test/AddAllExtensionsToApplicationArchiveProcessorTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package org.arquillian.container.liferay.remote.deploy.processor.test;
+package org.arquillian.container.osgi.remote.deploy.processor.test;
 
 import java.io.IOException;
 
@@ -24,13 +24,13 @@ import java.util.jar.Attributes;
 import java.util.jar.JarFile;
 import java.util.jar.Manifest;
 
-import org.arquillian.container.liferay.remote.activator.ArquillianBundleActivator;
-import org.arquillian.container.liferay.remote.deploy.processor.AddAllExtensionsToApplicationArchiveProcessor;
-import org.arquillian.container.liferay.remote.deploy.processor.test.mock.DummyInstanceProducerImpl;
-import org.arquillian.container.liferay.remote.deploy.processor.test.mock.DummyServiceLoaderWithJarAuxiliaryArchive;
-import org.arquillian.container.liferay.remote.deploy.processor.test.mock.DummyServiceLoaderWithOSGIBundleAuxiliaryArchive;
-import org.arquillian.container.liferay.remote.deploy.processor.test.mock.DummyServiceLoaderWithoutAuxiliaryArchive;
-import org.arquillian.container.liferay.remote.deploy.processor.test.util.ManifestUtil;
+import org.arquillian.container.osgi.remote.activator.ArquillianBundleActivator;
+import org.arquillian.container.osgi.remote.deploy.processor.test.mock.DummyInstanceProducerImpl;
+import org.arquillian.container.osgi.remote.deploy.processor.test.mock.DummyServiceLoaderWithJarAuxiliaryArchive;
+import org.arquillian.container.osgi.remote.deploy.processor.test.mock.DummyServiceLoaderWithOSGIBundleAuxiliaryArchive;
+import org.arquillian.container.osgi.remote.deploy.processor.test.mock.DummyServiceLoaderWithoutAuxiliaryArchive;
+import org.arquillian.container.osgi.remote.deploy.processor.test.util.ManifestUtil;
+import org.arquillian.container.osgi.remote.processor.AddAllExtensionsToApplicationArchiveProcessor;
 
 import org.jboss.arquillian.core.spi.ServiceLoader;
 import org.jboss.arquillian.test.spi.TestClass;
