+++ arquillian-container-osgi-with-extensions/src/test/java/test/org/arquillian/container/osgi/remote/deploy/processor/test/mock/DummyInstanceProducerImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package org.arquillian.container.liferay.remote.deploy.processor.test.mock;
+package org.arquillian.container.osgi.remote.deploy.processor.test.mock;
 
 import org.jboss.arquillian.core.api.InstanceProducer;
